Public Meeting View
New route for Public Meeting showing: /a/meetings/id/access_code.

Access code must be right to show any content.

Introduced new rails route to take id and code and return nothing or
the meeting if it’s valid code.

Public view is read-only, no interactions

Meeting views when logged in have a button to copy the public link to
clipboard.  Also restyled the meeting view a little to clean up.

Also, startMeeting button is “Resume” when the meeting is not unstated.